did the well structred workspace commit.
