+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "dotenv": "^17.1.0"
+  }
+}
