+++ simple-express-server/package.json
@@ -7,7 +7,8 @@
     "start": "nodemon src/server.js"
   },
   "dependencies": {
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "dotenv": "^16.4.5"
   },
   "devDependencies": {
     "nodemon": "^3.0.1"
