+++ simple-express-server/routes/root.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+// GET /
+router.get('/', (req, res) => {
+  res.send('Hello World');
+});
+
+module.exports = router;
